+++ formio/models/formio_version.py
@@ -1,6 +1,8 @@
 # Copyright Nova Code (http://www.novacode.nl)
 # See LICENSE file for full licensing details.
 
+from pkg_resources import parse_version
+
 from odoo import api, fields, models
 
 
@@ -52,7 +54,8 @@ class Version(models.Model):
     @api.model
     def _update_versions_sequence(self):
         versions = self.search([])
-        names = sorted(versions.mapped('name'))
+        names = versions.mapped('name')
+        names = sorted(names, key=parse_version)
         seq = 0
         for name in names:
             seq += 1